refactor: 프로덕션 코드 변경에 따른 테스트 수정
