+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,7 +3,6 @@ package kitchenpos.menu.application;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.menu.domain.MenuValidator;
 import kitchenpos.menu.ui.dto.request.MenuChangeRequest;
 import kitchenpos.menu.ui.dto.request.MenuRequest;
 import kitchenpos.menu.ui.dto.response.MenuResponse;
