--- src/test/java/kitchenpos/fixture/DefaultFixture.java
@@ -1,17 +1,0 @@
-package kitchenpos.fixture;
-
-import kitchenpos.testtool.RequestBuilder;
-import kitchenpos.testtool.request.RequestApi;
-
-public class DefaultFixture {
-
-    private final RequestBuilder requestBuilder;
-
-    public DefaultFixture(RequestBuilder requestBuilder) {
-        this.requestBuilder = requestBuilder;
-    }
-
-    protected RequestApi request() {
-        return requestBuilder.builder();
-    }
-}
